refactor(cef): split ui_tasks.rs into category submodules
