+++ .changesets/1782955411-refactor-cef-split-ui-tasks-into-category-submodules-1ddg.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(cef): split ui_tasks into category submodules
